Collateral case to fix XWM files stored as WAV
Some mod authors save XWM files with a WAV extension. This isn't an issue on PC as Skyrim will always check for the sound signature to use the correct codec. Added an additional check to verify if a WAV is indeed one. If a XWM is found it will convert to WAV before moving ahead. If anything else is found it won't create the DSP and return a false status from main conversion routine.